[Refactor] change package name
