+++ member-api/src/test/kotlin/io/waterkite94/stalk/api/controller/MemberRestControllerTest.kt
@@ -5,8 +5,8 @@ import io.waterkite94.stalk.api.dto.request.ChangeMemberPasswordRequest
 import io.waterkite94.stalk.api.dto.request.CreateMemberRequest
 import io.waterkite94.stalk.api.dto.request.UpdateMemberRequest
 import io.waterkite94.stalk.domain.model.Member
-import io.waterkite94.stalk.domain.vo.MemberStatus
-import io.waterkite94.stalk.domain.vo.RoleLevel
+import io.waterkite94.stalk.domain.type.MemberStatus
+import io.waterkite94.stalk.domain.type.RoleLevel
 import io.waterkite94.stalk.usecase.usecase.ChangeMemberProfile
 import io.waterkite94.stalk.usecase.usecase.CreateMember
 import io.waterkite94.stalk.usecase.usecase.VerifyEmail
