+++ member-domain/src/main/kotlin/io/waterkite94/stalk/domain/type/MemberStatus.kt
@@ -1,4 +1,4 @@
-package io.waterkite94.stalk.domain.vo
+package io.waterkite94.stalk.domain.type
 
 enum class MemberStatus {
     ACTIVE,
