+++ member-usecase/src/test/kotlin/io/waterkite94/stalk/usecase/usecase/ChangeMemberServiceTest.kt
@@ -2,8 +2,8 @@ package io.waterkite94.stalk.usecase.usecase
 
 import io.waterkite94.stalk.domain.model.Member
 import io.waterkite94.stalk.domain.model.UpdateMemberInformationDto
-import io.waterkite94.stalk.domain.vo.MemberStatus
-import io.waterkite94.stalk.domain.vo.RoleLevel
+import io.waterkite94.stalk.domain.type.MemberStatus
+import io.waterkite94.stalk.domain.type.RoleLevel
 import io.waterkite94.stalk.usecase.IntegrationTestSupport
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatThrownBy
